Specify provider and region for top level Terraform script
We need to specify that we want to deploy AWS infrastructure. So,
we create a provider block, and specify the region which to deploy
to. The region is set as a variable in a separate file in order to
provide a single point of control for a variable which often can
repeat.