Specify provider for new module
We want to create a module which will deploy an EC2 instance and
provision it to run the nginx Docker image. In order to do this
in an idiomatic way, we create a subdirectory in our Terraform
project called "modules", then create an additional subdirectory
for our specific module, which is called "demo".

This commit specifies the provider along with the region in the
same way as we did for the top level Terraform project.