Provision the EC2 instance
This commit provisions the EC2 instance, as well as the AMI that it
will be using. We arbitrarily choose a CentOS 8 AMI, and specify
the owner's account ID and image ID to make sure we are getting the
AMI that we want. We then create the EC2 resource and specify the
AMI that we provisioned. Additionally, this commit creates another
variable in the module which specifies the instance type. This
variable is used in the EC2 resource declaration.