Adds a provisioning script to the EC2 instance
In order to provision the EC2 instance, we need to add a user-data
script which will install Docker then run the nginx image. This
script is placed into a "templates" subdirectory in the module, then
templated alongside the EC2 instance, before being referenced by the
EC2 instance resource as a provisioning script.